refactor: fsd 공개 api와 세그먼트 구조 정리
슬라이스 루트 export를 보강하고 깊은 import를 줄였습니다.

허용되지 않은 세그먼트를 model/api 하위로 이동해 구조를 정리했습니다.

#61
