+++ apps/web/src/entities/alumni-contacts/model/index.ts
@@ -3,3 +3,7 @@ export {
   useAlumniContactsAcademicFilterSheetModalContext,
   useResetAlumniContactsFilter,
 } from './hooks';
+export type {
+  GetAlumniContactsQuery,
+  GetPaginatedAlumniContactsResponseDto,
+} from './types';
