+++ apps/web/src/entities/alumni-contacts/model/types/query.ts
