--- apps/web/src/entities/alumni-contacts/types/dto/index.ts
@@ -1,1 +1,0 @@
-export type { GetPaginatedAlumniContactsResponseDto } from './getAlumniContactsDto';
