--- apps/web/src/entities/alumni-contacts/types/query/index.ts
@@ -1,1 +1,0 @@
-export type { GetAlumniContactsQuery } from './getAlumniContactsQuery';
