+++ apps/web/src/entities/post/model/index.ts
@@ -1,3 +1,5 @@
 export * from './mock';
 export * from './types';
 export * from './form';
+export * from './queries/useGetPosts';
+export * from './queries/usePostQuery';
