+++ apps/web/src/features/auth/api/index.ts
@@ -1,1 +1,2 @@
 export * from './post';
+export * from './mock';
