+++ apps/web/src/widgets/alumni-contacts/ui/alumni-contacts-academic-filter-bottom-sheet/AlumniContactsAcademicFilterBottomSheet.tsx
@@ -1,6 +1,6 @@
 import { BottomSheet, CTAButton } from '@causw/cds';
 
-import { AlumniContactsAcademicFilterSheetModalMain } from '@/widgets/alumni-contacts/ui/alumni-contacts-academic-filter-sheet-modal-main';
+import { AlumniContactsAcademicFilterSheetModalMain } from '@/widgets/alumni-contacts';
 
 interface AlumniContactsAcademicFilterBottomSheetProps {
   open: boolean;
