+++ apps/web/src/widgets/alumni-contacts/ui/index.ts
@@ -3,3 +3,4 @@ export {
   AlumniContactsListServerComponent,
   AlumniContactsListLoadingView,
 } from './alumni-contacts-list';
+export { AlumniContactsAcademicFilterSheetModalMain } from './alumni-contacts-academic-filter-sheet-modal-main';
