+++ apps/web/src/widgets/setting/index.ts
@@ -1,2 +1,3 @@
-﻿export * from './model';
+export * from './model';
 export * from './ui';
+export * from './config';
